training process has started !

+++ running/util.py
@@ -1,6 +1,5 @@
 import numpy as np
 import torch
-import torch.nn.functional as F
 
 
 # Based on https://github.com/DongjunLee/transformer-tensorflow/blob/master/transformer/attention.py
@@ -19,7 +18,7 @@ def prefix_sum(arr):
 
 
 def sparse_categorical_accuracy(y_true, y_pred):
-    return torch.eq(y_true, torch.argmax(y_pred, -1))
+    return torch.eq(y_true, torch.argmax(y_pred, -1)).long()
 
 
 def sparse_softmax_cross_entropy_with_logits(error_locations, loc_predictions):
